+++ resources/views/admin/posts/show.blade.php
@@ -5,7 +5,7 @@
 	<x-row flex-class="md:flex flex-col gap-6">
 		<x-column class="block w-full">
 			@if($post->image)
-				<figure class="lightbox-trigger">
+				<figure class="lightbox-trigger inline-block">
 					<img src="{{ $post->image->url }}" class="block rounded-lg" alt="">
 				</figure>
 			@endif